refactor(contexts): consolidate TranslationsContext into contexts directory
Move the TranslationsContext and related imports from the components directory to a new contexts directory to improve code organization and maintainability. This change centralizes context-related files and reduces redundancy in import paths.
