+++ src/components/Settings/AutoStart.tsx
@@ -4,7 +4,7 @@ import { createSignal, onMount } from "solid-js";
 import { checkAutoStart, disableAutoStart, enableAutoStart } from "~/commands";
 import { message } from "@tauri-apps/plugin-dialog";
 
-import { useTranslations } from "../TranslationsContext";
+import { useTranslations } from "~/contexts";
 
 const AutoStart = () => {
   const { translate, translateErrorMessage } = useTranslations();
